Placed FFs Side by Side and Discussed Table Metrics
Changes the FF plots with and without the chi^2 cut so that they are directly above one another for easier comparison. Also wrote a bit explaining the table metrics like the sum of the Qi parameters. Explained when we didn't force the Qi to sum to unity and why we didn't float the normalization of the individual data sets.