Lots of Final Revisions
Fixed the red line for my fits in the legends. Made the Qi and Ri plots slightly more readable. More discussion of the form factors and charge radii was added to chapter one. Discussed the asymmetry measurement experiment briefly to end the conclusions section. Added a zoomed in plot of my xs fit with my point and it's error bars that had been neglected.