Continuing Revisions Charge Radii
Improved the discussion of the charge radii and included the additional uncertainty due to the sum Qi not being forced to unity. Hopefully clear now that the standard deviations I quote underestimate the uncertainty in the charge radii (probably not much for 3He and probably a significant amount for 3H).